+++ omnisafe/algos/model_based/policy_gradient.py
@@ -24,9 +24,9 @@ from torch.optim import Adam
 from omnisafe.algos import registry
 from omnisafe.algos.common.logger import Logger
 from omnisafe.algos.common.replay_buffer import ReplayBuffer as Off_ReplayBuffer
-from omnisafe.algos.model_based.models.core_ac import MLPActorCritic
-from omnisafe.algos.model_based.models.dynamics_predict_env import PredictEnv
-from omnisafe.algos.model_based.models.dynamicsmodel import EnsembleDynamicsModel
+from omnisafe.algos.model_based.models.actor_critic import MLPActorCritic
+from omnisafe.algos.model_based.models.virtual_env import VirtualEnv
+from omnisafe.algos.model_based.models.dynamic_model import EnsembleDynamicsModel
 from omnisafe.algos.utils.distributed_utils import proc_id
 
 
@@ -72,7 +72,7 @@ class PolicyGradientModelBased:  # pylint: disable=too-many-instance-attributes,
             cost_size=0,
             **self.cfgs['dynamics_cfgs'],
         )
-        self.predict_env = PredictEnv(algo, self.dynamics, self.env_id, self.device)
+        self.predict_env = VirtualEnv(algo, self.dynamics, self.env_id, self.device)
 
         # Initialize off-policy buffer
         # pylint: disable-next=line-too-long
